+++ earthdaily/__init__.py
@@ -1,3 +1,3 @@
 from . import earthdatastore, datasets
 
-__version__ = "0.0.5"
+__version__ = "0.0.6"