+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2024-02-23
+
+### Fixed
+
+- `to_wkt`for GeometryManager.
+
+### Added
+
+- Add docstring thanks to @piclem.
+- Token support thanks to @luisageo6.
+
 ## [0.0.5] - 2024-02-19
 
 ### Fixed